Move ContentContainer into contentContainer folder
